Add 'chalice dev appgraph' command
